moving index back to root as per code review comments
